+++ plumage/__init__.py
@@ -20,6 +20,8 @@
 from pathlib import Path
 
 import pelican
+from pelican import signals
+from pelican.contents import Static
 
 from .dom_transforms import transform
 
@@ -67,6 +69,38 @@ def check_config(sender):
             f"{code_style} not recognized among {sorted(ALL_CODE_STYLES)}."
         )
 
+def add_favicon_assets(sender):
+    """ Copy all individual files found in /static/favicon theme's folder
+    to the root of the generated site.
 
-pelican.signals.initialized.connect(check_config)
-pelican.signals.content_written.connect(transform)
+    Favicons were generated with RealFaviconGenerator v0.16:
+    https://realfavicongenerator.net
+    """
+    # Our theme structure expect default configuration.
+    # See: https://docs.getpelican.com/en/latest/settings.html#themes
+    assert "static" in sender.settings["THEME_STATIC_PATHS"]
+
+    for asset in (PLUMAGE_ROOT / 'static' / 'favicon').iterdir():
+        assert asset.is_file()
+
+        # Add static file to generator context, as Pelican does natively at:
+        # https://github.com/getpelican/pelican/blob/8033162ba4393db60791b201fb100d1be0f04431/pelican/generators.py#L811-L817
+        static = sender.readers.read_file(
+                base_path=asset.parent, path=asset.name, content_class=Static,
+                fmt='static', context=sender.context,
+                preread_signal=signals.static_generator_preread,
+                preread_sender=sender,
+                context_signal=signals.static_generator_context,
+                context_sender=sender)
+
+        # Forces the asset to be saved at the root of the output folder.
+        # See: https://github.com/getpelican/pelican/blob/8033162ba4393db60791b201fb100d1be0f04431/pelican/contents.py#L55-L59
+        static.metadata['save_as'] = asset.name
+        setattr(static, 'override_save_as', asset.name)
+
+        # Register asset along other static files.
+        sender.context['staticfiles'].append(static)
+
+signals.initialized.connect(check_config)
+signals.static_generator_finalized.connect(add_favicon_assets)
+signals.content_written.connect(transform)